reorganize scripts into folders and namespaces
