--- Assets/Level/Scenes/Sandbox Scenes/PlatformerPlusPlayer.unity.orig.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: fcfd17309a0f0b944a74fc0abf4d0eb4
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
